refactor: move chat components from global components/ to views/chat/components/
 - Move ChatInput, ChatMessage, SessionList, TracePanel from src/components/chat/ to src/views/chat/components/
 - Update import paths in ChatView.vue to use the new location
 - Delete the now-empty src/components/chat/ directory
