+++ data-agent-frontend/src/views/chat/components/ChatInput.vue
