+++ data-agent-frontend/src/views/chat/components/SessionList.vue
