+++ data-agent-frontend/src/views/chat/components/TracePanel.vue
